jdbc/src/test/java/com/epam/trainings/java/fundamentals/jdbc/cp/ConnectionPoolTest.java
package com.epam.trainings.java.fundamentals.jdbc.cp;

import org.junit.BeforeClass;
import org.junit.Test;

import java.sql.Connection;
import java.sql.ResultSet;
import java.sql.Statement;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;
import static org.junit.Assert.assertTrue;

public class ConnectionPoolTest {

    private static final String RESOURCES_FILE_PATH = "src/test/resources/";
    private static final String DB_PROPERTIES_FILE_NAME = "db.properties";
    private static final String DB_PREPARE_FILE_NAME = "h2.sql";

    static ConnectionPool connectionPool;

    @BeforeClass
    public static void init() throws ConnectionPoolException {
        connectionPool = ConnectionPool.create(RESOURCES_FILE_PATH + DB_PROPERTIES_FILE_NAME);
        connectionPool.executeScript(RESOURCES_FILE_PATH + DB_PREPARE_FILE_NAME);
    }

    @Test
    public void takeConnection() throws Exception {
        try (Connection connection = connectionPool.getConnection();
             Statement statement = connection.createStatement();
             ResultSet resultSet = statement.executeQuery("SELECT name FROM Gun WHERE id = 1")) {
            assertTrue(resultSet.next());
            assertThat(resultSet.getString(1), is("Kolt"));
        }
    }
}